Add support for multi-channel power supplies
We do this by adding a `channel` argument to all the methods that set or
get the state of an output. The underlying assumption here is that the
outputs can be adjusted independently but are all enabled together.

We then create classes that exposes a `PowerSupplyChannel` for each
channel they support.