Add factory method to create `PowerSupply`
Since we now have three different `PowerSupply` classes, we want a
factory method that will create the appropriate one for us.

The number of channels in a power supply is know ahead of time so we can
store it in the `PowerSupplyDescription` and use that to return the
correct class.

For this, we add a `make_power_supply` function that returns an instance
of `PowerSupply` (or one of the daughter classes) based on the identity
of the power supply.